% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portawaterperu.R
\docType{data}
\name{portawaterperu}
\alias{portawaterperu}
\title{portawaterperu: A Preliminary Review of Peruvian Potable Water System Data}
\format{
A tibble with 6 rows and 18 variables
\describe{
\item{nombre}{name of community water system}
\item{ID}{identification number (unique to each water system)}
\item{pais}{Country of data collection}
\item{divisiones}{geographic division (region; province; district)}
\item{latitud}{latitude of district}
\item{longitud}{longitude of district}
\item{altitud}{altitude of district}
\item{ano}{year water system data was collected}
\item{comunidades}{community/municipality/town name}
\item{PSE}{water service provider}
\item{pob_servida}{population number served by water service provider}
\item{viv_servidas}{household number served by water service provider}
\item{tipo_gravedad}{is the community served by a gravity distribution system?}
\item{tipo_bombeo}{is the community served by a force/pump distribution system}
\item{tipo_pozo_manual}{is the community served by a well system (communal collection point / no distribution pipes)}
\item{tipo_agua_lluvia}{rain water (not sure in what regard this relates to water system won't use in analysis)}
\item{agua_epoca_seca}{dry season water (not sure in what regard this relates to water system won't use in analysis)}
\item{agua_epoca_lluvia}{rainy season water (not sure in what regard this relates to water system won't use in analysis)}
}
}
\usage{
portawaterperu
}
\description{
Data for this project was obtained from @SIASAR, an information system containing data on rural water supply and sanitation services. Using SIASAR's "download data by country" tool, all available data for Peru (10 excel files) were downloaded. After examining the 10 excel files, only 5 pertained to potable water systems. Those 5 data sets were imported into R and subsequently empty values and unnecessary columns were deleted from them. Finally, the 5 data sets were combined into 1 data frame based on community ID. The combined, cleaned data set contains data from 32 communities.
}
\keyword{datasets}
